Split ConfigConvert object into multiple traits
